WIP data file for CMSC students
ClassData2.js is a work in progress (WIP) draft of some the classes that at least need consideration based on the sample1.html file for prerequisites and courses that CMSC students could take on their path to graduation. One path for SCIENCE is taken into consideration, but in the future, additional SCIENCE paths may be added to the data file if deemed appropriate. 

Once we decide on how to format the data file, more permanent changes can be made to it. 

Questions will be asked and answered during the Google Hangout section.