Description added to all courses in classdata.json file, some progress in using code to accomplish this
Changes made to the copy of ClassDatabase.js (i.e.
ClassDatabaseCopy.js) in terms of modifying the constructor to
accommodate a description, adding the uploadDataFile function from
Project 1 so that ideally another .json file would be created populated
with the additional description for each course in the classmate.json
file, and some work added to the addNewClassData function as a start
towards having this function contribute to the process of having the
course descriptions for each class added to the classdata.json file.
Variable des also added to the createClassFromRawData function to
account for desire to have each class have its description added to the
classdata.json file.

All modifications made to the ClassDatabaseCopy.js file so as to
perserve the integrity of the original ClassDatabaseCopy.js file, at
least until it has all the proper changes to completely have the code
in it we need, possibly from the ClassDatabaseCopy.js file if time
permits.